+++ substrate/client/consensus/aura/src/lib.rs
@@ -60,13 +60,14 @@ pub use import_queue::{
 	build_verifier, import_queue, AuraVerifier, BuildVerifierParams, CheckForEquivocation,
 	ImportQueueParams,
 };
-pub use sc_consensus_slots::{InherentDigest, SlotProportion};
+pub use sc_consensus_slots::SlotProportion;
 pub use sp_consensus::SyncOracle;
 pub use sp_consensus_aura::{
 	digests::CompatibleDigestItem,
 	inherents::{InherentDataProvider, InherentType as AuraInherent, INHERENT_IDENTIFIER},
 	AuraApi, ConsensusLog, SlotDuration, AURA_ENGINE_ID,
 };
+pub use sp_consensus_slots::InherentDigest;
 
 const LOG_TARGET: &str = "aura";
