+++ substrate/primitives/consensus/slots/Cargo.toml
@@ -19,6 +19,8 @@ targets = ["x86_64-unknown-linux-gnu"]
 codec = { package = "parity-scale-codec", version = "3.6.1", default-features = false, features = ["derive", "max-encoded-len"] }
 scale-info = { version = "2.11.1", default-features = false, features = ["derive"] }
 serde = { features = ["alloc", "derive"], optional = true, workspace = true }
+sp-inherents = { path = "../../inherents", default-features = false }
+sp-runtime = { path = "../../runtime", optional = true, default-features = false }
 sp-timestamp = { path = "../../timestamp", default-features = false }
 
 [features]
@@ -27,6 +29,8 @@ std = [
 	"codec/std",
 	"scale-info/std",
 	"serde/std",
+	"sp-inherents/std",
+	"sp-runtime/std",
 	"sp-timestamp/std",
 ]
